dataset: all returned dataset must be follow the same mode with original.
When splitting and random pick the dataset, the returned dataset must
have the same mode with original dataset. For example, if original dataset
D mode is "matrix" and the function call split on D, the returned value
has the same mode, which is also "matrix", not "rows".

This commit also fix some transpose and push issues,

* add new function to push column to rows.
* PushRow: check the mode when pushing new row.
* TransposeToRows: only reset columns if origin mode is column, this
prevent empty columns if mode is "matrix".
* TransposeToColumns: only reset rows if origin mode is rows, this
prevent empty rows if if mode is "matrix".
* GetNRow: fix returned value if mode is matrix.